// CBHIS-Fontend-stable-training-v1.1.1/mobile-apps/src/app/modules/client/enums/malaria.enum.ts
export enum ISR {
  ISR = 1,
}

export enum NumberOfITN {
  NumberOfITN = 1,
}

export enum ISRProvider {
  GovernmentWorkerProgram = 1,
  PrivateCompany = 2,
  NGO = 3,
  Unknown = 4,
}

export enum MalariaCampaign {
  Radio = 1,
  Television = 2,
  Poster = 3,
  Billboard = 4,
  RHM = 5,
  CommunityEvent = 6,
  NGO = 7,
}